requirements/base.txt
django==1.9.5
django-authtools
django-extensions
django-grappelli
django-redis
django-redis-cache
factory-boy
fake-factory
psycopg2
purl
uwsgi
bleach
boto
dj-database-url
dj-static
django-embed-video
django-guardian
django-storages
easy-thumbnails
gunicorn
html5lib
Markdown
Pillow==2.7.0 
python-decouple
requests
six
static3
Unipath
whitenoise
httplib2
oauth2
python-openid
django-annoying
python-social-auth
django-ckeditor
django-mptt
pathlib
